+++ ADC Installer/src/com/kaylerrenslow/armaDialogCreator/installer/ADCInstaller.java
@@ -1,10 +1,12 @@
 package com.kaylerrenslow.armaDialogCreator.installer;
 
 import javafx.application.Application;
-import javafx.application.Platform;
 import javafx.stage.Stage;
+import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.util.ResourceBundle;
 
 /**
@@ -19,24 +21,29 @@ public class ADCInstaller extends Application {
 
 	@Override
 	public void start(Stage primaryStage) throws Exception {
-		File f = new File("");
-		if (this.getParameters().getUnnamed().contains("-z")) {
-			String zipFilePath = this.getParameters().getNamed().get("f");
-			if (zipFilePath == null) {
-				System.out.println("f is missing. Expected --f=zipFile");
-				Platform.exit();
-				return;
-			}
-			File zipFile = new File(zipFilePath);
-
+		File f = new File(System.getProperty("user.home") + "/Arma Dialog Creator");
+		if (!f.exists()) {
+			f.mkdirs();
+		} else if (f.exists() && !f.isDirectory()) {
+			f = new File(System.getProperty("user.home"));
+		}
+		if (this.getParameters().getUnnamed().contains("-noui")) {
 			InstallerProgressWindow pw = new InstallerProgressWindow(primaryStage);
 			primaryStage.show();
-			pw.runInstall(zipFile, f);
+			pw.runInstall(f);
 		} else {
 			InstallerConfigWindow window = new InstallerConfigWindow(primaryStage, f);
 			primaryStage.show();
 		}
+	}
 
+	@NotNull
+	public static String getExceptionString(@NotNull Throwable t) {
+		StringWriter sw = new StringWriter();
+		PrintWriter pw = new PrintWriter(sw);
+		t.printStackTrace(pw);
+		pw.close();
+		return sw.toString();
 	}
 
 }