+++ CHANGES.md
@@ -1,8 +1,9 @@
 **Added:**
-* 
+* message for install config window when install failed
 
 **Changed:**
-* made some progress on installer again
+* removed support for installer using .zip (will now only extract out of installer jar)
+* folder where arma dialog creator saves config stuff
 
 **Fixed:**
 * none